Incorpoate new code and refactor to run more functions as airflow tasks
