+++ academic_observatory_workflows/workflows/oa_web_workflow.py
@@ -25,7 +25,6 @@ import os
 import os.path
 import shutil
 import statistics
-import urllib.parse
 from concurrent.futures import ThreadPoolExecutor, as_completed
 from dataclasses import field
 from typing import List, Dict, Set
@@ -39,7 +38,6 @@ import nltk
 import numpy as np
 import pandas as pd
 import pendulum
-import requests
 import treelib
 from airflow.exceptions import AirflowException
 from airflow.models.variable import Variable
@@ -49,9 +47,12 @@ from pandas.api.types import is_string_dtype
 
 from academic_observatory_workflows.clearbit import clearbit_download_logo
 from academic_observatory_workflows.dag_tag import Tag
+from academic_observatory_workflows.github import trigger_repository_dispatch
+from academic_observatory_workflows.wikipedia import fetch_wiki_descriptions
+from academic_observatory_workflows.zenodo import Zenodo, make_draft_version, publish_new_version
 from observatory.platform.utils.airflow_utils import AirflowVars, get_airflow_connection_password
 from observatory.platform.utils.config_utils import module_file_path
-from observatory.platform.utils.file_utils import load_jsonl
+from observatory.platform.utils.file_utils import load_jsonl, list_to_jsonl_gz
 from observatory.platform.utils.gc_utils import (
     bigquery_sharded_table_id,
     download_blobs_from_cloud_storage,
@@ -64,24 +65,64 @@ from observatory.platform.workflows.snapshot_telescope import SnapshotRelease
 from observatory.platform.workflows.workflow import Workflow
 
 # The minimum number of outputs before including an entity in the analysis
-INCLUSION_THRESHOLD = {"country": 25, "institution": 350}
+PERCENTAGE_FIELD_KEYS = [
+    ("outputs_open", "n_outputs"),
+    ("outputs_both", "n_outputs"),
+    ("outputs_closed", "n_outputs"),
+    ("outputs_publisher_open", "n_outputs"),
+    ("outputs_publisher_open_only", "n_outputs"),
+    ("outputs_other_platform_open", "n_outputs"),
+    ("outputs_other_platform_open_only", "n_outputs"),
+    ("outputs_oa_journal", "n_outputs_publisher_open"),
+    ("outputs_hybrid", "n_outputs_publisher_open"),
+    ("outputs_no_guarantees", "n_outputs_publisher_open"),
+    ("outputs_preprint", "n_outputs_other_platform_open"),
+    ("outputs_domain", "n_outputs_other_platform_open"),
+    ("outputs_institution", "n_outputs_other_platform_open"),
+    ("outputs_public", "n_outputs_other_platform_open"),
+    ("outputs_other_internet", "n_outputs_other_platform_open"),
+]
+INCLUSION_THRESHOLD = {"country": 1, "institution": 1}
 MAX_REPOSITORIES = 200
 START_YEAR = 2000
+END_YEAR = pendulum.now().year - 1
+WIKI_MAX_TITLES = 20  # Set the number of titles for which wiki descriptions are retrieved at once, the API can return max 20 extracts.
 
-# The query that pulls data to be included in the dashboards
-QUERY = """
+# Queries that pull data
+INSTITUTION_INDEX_QUERY = """
+SELECT
+  ror.id,
+  ror.name,
+  (SELECT * from ror.links LIMIT 1) AS url,
+  ror.wikipedia_url,
+  country.wikipedia_name as country,
+  country.subregion as subregion,
+  country.region as region,
+  ror.types AS institution_types,
+  ror.external_ids AS identifiers,
+  ror.acronyms
+FROM
+  `{project_id}.{ror_dataset_id}.{ror_table_id}` as ror
+  LEFT OUTER JOIN `{project_id}.{settings_dataset_id}.{country_table_id}` as country ON ror.country.country_code = country.alpha2
+ORDER BY name ASC
+"""
+
+COUNTRY_INDEX_QUERY = """
+SELECT
+  country.wikipedia_name as id,
+  country.wikipedia_name as name,
+  country.wikipedia_url,
+  country.subregion as subregion,
+  country.region as region,
+FROM
+  `{project_id}.{settings_dataset_id}.{country_table_id}` as country
+ORDER BY name ASC
+"""
+
+DATA_QUERY = """
 SELECT
   agg.id,
-  agg.name,
   agg.time_period as year,
-  DATE(agg.time_period, 12, 31) as date,
-  (SELECT * from ror.links LIMIT 1) AS url,
-  COALESCE(ror.wikipedia_url, country.wikipedia_url) as wikipedia_url,
-  country.alpha2 as alpha2,
-  agg.country as country,
-  agg.subregion as subregion,
-  agg.region as region,
-  ror.types AS institution_types,  
   agg.citations.mag.total_citations as n_citations,  
   agg.total_outputs as n_outputs,
   
@@ -105,19 +146,12 @@ SELECT
   agg.oa_coki.other_platform_categories.institution.total AS n_outputs_institution,
   agg.oa_coki.other_platform_categories.public.total AS n_outputs_public,
   agg.oa_coki.other_platform_categories.aggregator.total + agg.oa_coki.other_platform_categories.other_internet.total + agg.oa_coki.other_platform_categories.unknown.total AS n_outputs_other_internet, 
-  
-  ror.external_ids AS identifiers,
+
   agg.repositories,
-  CASE
-    WHEN agg.id = country.alpha3 THEN []
-    ELSE ror.acronyms
-    END AS acronyms,
 FROM
-  `{project_id}.{agg_dataset_id}.{agg_table_id}` as agg 
-  LEFT OUTER JOIN `{project_id}.{ror_dataset_id}.{ror_table_id}` as ror ON agg.id = ror.id
-  LEFT OUTER JOIN `{project_id}.{settings_dataset_id}.{country_table_id}` as country ON agg.id = country.alpha3
-WHERE agg.time_period >= {start_year} AND agg.time_period <= (EXTRACT(YEAR FROM CURRENT_DATE()) - 1)
-ORDER BY year DESC, name ASC
+  `{project_id}.{agg_dataset_id}.{agg_table_id}` as agg
+WHERE agg.time_period >= 2000 AND agg.time_period <= {end_year}
+ORDER BY year DESC
 """
 
 README = """# COKI Open Access Dataset
@@ -144,6 +178,644 @@ The COKI Open Access Dataset contains information from:
 * [Unpaywall](https://unpaywall.org/). The [Unpaywall Data Feed](https://unpaywall.org/products/data-feed) is used under license. Data is freely available from Unpaywall via the API, data dumps and as a data feed.
 * [Research Organization Registry](https://ror.org/) which is made available under a [CC0 licence](https://creativecommons.org/share-your-work/public-domain/cc0/).
 """
+
+###################
+# Airflow Workflow
+###################
+
+
+class OaWebRelease(SnapshotRelease):
+    def __init__(
+        self,
+        *,
+        dag_id: str,
+        release_date: pendulum.DateTime,
+        data_bucket_name: str,
+        zenodo: Zenodo = Zenodo(),
+    ):
+        """Create an OaWebRelease instance.
+
+        :param dag_id: the dag id.
+        :param release_date: the release date.
+        :param zenodo: the zenodo instance.
+        """
+
+        super().__init__(dag_id=dag_id, release_date=release_date)
+        self.zenodo = zenodo
+        self.data_bucket_name = data_bucket_name
+        self.assets_path = module_file_path("academic_observatory_workflows.workflows.data.oa_web_workflow")
+
+    @property
+    def build_path(self):
+        path = os.path.join(self.transform_folder, "build")
+        os.makedirs(path, exist_ok=True)
+        return path
+
+    @property
+    def intermediate_path(self):
+        path = os.path.join(self.transform_folder, "intermediate")
+        os.makedirs(path, exist_ok=True)
+        return path
+
+
+class OaWebWorkflow(Workflow):
+    DATA_BUCKET = "oa_web_data_bucket"
+    GITHUB_TOKEN_CONN = "oa_web_github_token"
+    ZENODO_TOKEN_CONN = "oa_web_zenodo_token"
+    ROR_FILE = "ror.jsonl.gz"
+    COUNTRY_INDEX_FILE = "country-index.jsonl.gz"
+    INSTITUTION_INDEX_FILE = "institution-index.jsonl.gz"
+    COUNTRY_DATA_FILE = "country-data.jsonl.gz"
+    INSTITUTION_DATA_FILE = "institution-data.jsonl.gz"
+
+    """The OaWebWorkflow generates data files for the COKI Open Access Dashboard.
+
+    The figure below illustrates the generated data and notes about what each file is used for.
+    .
+    ├── data: data
+    │   ├── index.json: used by the Cloudflare Worker search and filtering API.
+    │   ├── autocomplete.json: used for the website search functionality. Copied into public/data folder.
+    │   ├── country: individual entity statistics files for countries. Used to build each country page.
+    │   │   ├── ALB.json
+    │   │   ├── ARE.json
+    │   │   └── ARG.json
+    │   ├── country.json: used to create the country table. First 18 countries used to build first page of country table
+    │   │                 and then this file is included in the public folder and downloaded by the client to enable the
+    │   │                 other pages of the table to be displayed. Copied into public/data folder.
+    │   ├── institution: individual entity statistics files for institutions. Used to build each institution page.
+    │   │   ├── 05ykr0121.json
+    │   │   ├── 05ym42410.json
+    │   │   └── 05ynxx418.json
+    │   ├── institution.json: used to create the institution table. First 18 institutions used to build first page of institution table
+    │   │                     and then this file is included in the public folder and downloaded by the client to enable the
+    │   │                     other pages of the table to be displayed. Copied into public/data folder.
+    │   └── stats.json: global statistics, e.g. the minimum and maximum date for the dataset, when it was last updated etc.
+    └── logos: country and institution logos. Copied into public/logos folder.
+        ├── country
+        │   ├── l: large logos displayed on country pages.
+        │   │   ├── ALB.svg
+        │   │   ├── ARE.svg
+        │   │   └── ARG.svg
+        │   └── s: small logos displayed in country table.
+        │       ├── ALB.svg
+        │       ├── ARE.svg
+        │       └── ARG.svg
+        └── institution
+            ├── l: large logos displayed on institution pages.
+            │   ├── 05ykr0121.jpg
+            │   ├── 05ym42410.jpg
+            │   └── 05ynxx418.jpg
+            └── s: small logos displayed in institution table.
+                ├── 05ykr0121.jpg
+                ├── 05ym42410.jpg
+                └── 05ynxx418.jpg
+    """
+
+    def __init__(
+        self,
+        *,
+        input_project_id: str = "academic-observatory",
+        output_project_id: str = "academic-observatory",
+        dag_id: str = "oa_web_workflow",
+        start_date: Optional[pendulum.DateTime] = pendulum.datetime(2021, 5, 2),
+        schedule_interval: Optional[str] = "@weekly",
+        catchup: Optional[bool] = False,
+        ext_dag_id: str = "doi",
+        table_ids: List[str] = None,
+        airflow_vars: List[str] = None,
+        airflow_conns: List[str] = None,
+        agg_dataset_id: str = "observatory",
+        ror_dataset_id: str = "ror",
+        settings_dataset_id: str = "settings",
+        version: str = "v6",
+        conceptrecid: int = 6399462,
+        zenodo_host: str = "https://zenodo.org",
+    ):
+        """Create the OaWebWorkflow.
+
+        :param project_id: the Google Cloud project id.
+        :param dag_id: the DAG id.
+        :param start_date: the start date.
+        :param schedule_interval: the schedule interval.
+        :param catchup: whether to catchup or not.
+        :param ext_dag_id: the DAG id to wait for.
+        :param table_ids: the table ids.
+        :param airflow_vars: required Airflow Variables.
+        :param airflow_conns: required Airflow Connections.
+        :param agg_dataset_id: the id of the dataset where the Academic Observatory aggregated data lives.
+        :param ror_dataset_id: the id of the dataset containing the ROR table.
+        :param settings_dataset_id: the id of the settings dataset, which contains the country table.
+        :param version: the dataset version published by this workflow. The Github Action pulls from a specific dataset
+        version: https://github.com/The-Academic-Observatory/coki-oa-web/blob/develop/.github/workflows/build-on-data-update.yml#L68-L74.
+        This is so that when breaking changes are made to the schema, the web application won't break.
+        :param conceptrecid: the Zenodo Concept Record ID for the COKI Open Access Dataset. The Concept Record ID is
+        the last set of numbers from the Concept DOI.
+        :param zenodo_host: the Zenodo hostname, can be changed to https://sandbox.zenodo.org for testing.
+        """
+
+        if airflow_vars is None:
+            airflow_vars = [
+                AirflowVars.DATA_PATH,
+                AirflowVars.PROJECT_ID,
+                AirflowVars.DATA_LOCATION,
+                AirflowVars.DOWNLOAD_BUCKET,
+                AirflowVars.TRANSFORM_BUCKET,
+                self.DATA_BUCKET,
+            ]
+
+        if airflow_conns is None:
+            airflow_conns = [self.GITHUB_TOKEN_CONN, self.ZENODO_TOKEN_CONN]
+
+        super().__init__(
+            dag_id=dag_id,
+            start_date=start_date,
+            schedule_interval=schedule_interval,
+            catchup=catchup,
+            airflow_vars=airflow_vars,
+            airflow_conns=airflow_conns,
+            tags=[Tag.academic_observatory],
+        )
+        self.input_project_id = input_project_id
+        self.output_project_id = output_project_id
+        self.agg_dataset_id = agg_dataset_id
+        self.ror_dataset_id = ror_dataset_id
+        self.settings_dataset_id = settings_dataset_id
+        self.table_ids = table_ids
+        self.version = version
+        self.conceptrecid = conceptrecid
+        self.zenodo_host = zenodo_host
+        if table_ids is None:
+            self.table_ids = ["country", "institution"]
+
+        self.add_operator(
+            ExternalTaskSensor(task_id=f"{ext_dag_id}_sensor", external_dag_id=ext_dag_id, mode="reschedule")
+        )
+        self.add_setup_task(self.check_dependencies)
+        self.add_task(self.query)
+        self.add_task(self.download)
+        self.add_task(self.make_draft_zenodo_version)
+        self.add_task(self.download_twitter_cards)
+        self.add_task(self.preprocess_data)
+        self.add_task(self.build_indexes)
+        self.add_task(self.download_logos)
+        self.add_task(self.download_wiki_descriptions)
+        self.add_task(self.build_datasets)
+        self.add_task(self.publish_zenodo_version)
+        self.add_task(self.upload_dataset)
+        self.add_task(self.repository_dispatch)
+        self.add_task(self.cleanup)
+
+    ######################################
+    # Airflow tasks
+    ######################################
+
+    def make_release(self, **kwargs) -> OaWebRelease:
+        """Make release instances. The release is passed as an argument to the function (TelescopeFunction) that is
+        called in 'task_callable'.
+
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: A list of OaWebRelease instances
+        """
+
+        release_date = make_release_date(**kwargs)
+        data_bucket_name = Variable.get(self.DATA_BUCKET)
+        zenodo_token = get_airflow_connection_password(self.ZENODO_TOKEN_CONN)
+        zenodo = Zenodo(host=self.zenodo_host, access_token=zenodo_token)
+
+        return OaWebRelease(
+            dag_id=self.dag_id,
+            data_bucket_name=data_bucket_name,
+            release_date=release_date,
+            zenodo=zenodo,
+        )
+
+    def query(self, release: OaWebRelease, **kwargs):
+        """Fetch the data for each table.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        results = []
+        queries = []
+
+        # Query the country and institution aggregations
+        for agg_table_id in self.table_ids:
+            # Aggregate release dates
+            agg_release_date = select_table_shard_dates(
+                project_id=self.input_project_id,
+                dataset_id=self.agg_dataset_id,
+                table_id=agg_table_id,
+                end_date=release.release_date,
+            )[0]
+            agg_sharded_table_id = bigquery_sharded_table_id(agg_table_id, agg_release_date)
+
+            # Fetch data
+            destination_uri = (
+                f"gs://{release.download_bucket}/{self.dag_id}/{release.release_id}/{agg_table_id}-data.jsonl.gz"
+            )
+            query = (
+                DATA_QUERY.format(
+                    start_year=START_YEAR,
+                    project_id=self.input_project_id,
+                    agg_dataset_id=self.agg_dataset_id,
+                    agg_table_id=agg_sharded_table_id,
+                    end_year=END_YEAR,
+                ),
+            )
+            queries.append((query, destination_uri))
+
+        # Query ROR table
+        ror_table_id = "ror"
+        ror_release_date = select_table_shard_dates(
+            project_id=self.input_project_id,
+            dataset_id=self.ror_dataset_id,
+            table_id=ror_table_id,
+            end_date=release.release_date,
+        )[0]
+        ror_sharded_table_id = bigquery_sharded_table_id(ror_table_id, ror_release_date)
+        query = f"SELECT * FROM {self.input_project_id}.{self.ror_dataset_id}.{ror_sharded_table_id}"
+        destination_uri = f"gs://{release.download_bucket}/{self.dag_id}/{release.release_id}/{self.ROR_FILE}"
+        queries.append((query, destination_uri))
+
+        # Query institution index table
+        destination_uri = (
+            f"gs://{release.download_bucket}/{self.dag_id}/{release.release_id}/{self.INSTITUTION_INDEX_FILE}"
+        )
+        queries.append(
+            (
+                INSTITUTION_INDEX_QUERY.format(
+                    project_id=self.input_project_id,
+                    ror_dataset_id=self.ror_dataset_id,
+                    ror_table_id=ror_sharded_table_id,
+                    settings_dataset_id=self.settings_dataset_id,
+                    country_table_id="country",
+                ),
+                destination_uri,
+            )
+        )
+
+        # Query the country index table
+        destination_uri = f"gs://{release.download_bucket}/{self.dag_id}/{release.release_id}/{self.COUNTRY_INDEX_FILE}"
+        queries.append(
+            (
+                COUNTRY_INDEX_QUERY.format(
+                    project_id=self.input_project_id,
+                    settings_dataset_id=self.settings_dataset_id,
+                    country_table_id="country",
+                ),
+                destination_uri,
+            )
+        )
+
+        # Run queries, saving results to Google Cloud Storage
+        for (query, destination_uri) in queries:
+            success = bq_query_to_gcs(
+                query=query,
+                project_id=self.output_project_id,
+                destination_uri=destination_uri,
+            )
+            results.append(success)
+
+        state = all(results)
+        if not state:
+            raise AirflowException("OaWebWorkflow.query failed")
+
+    def download(self, release: OaWebRelease, **kwargs):
+        """Download the queried data.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        prefix = f"{self.dag_id}/{release.release_id}"
+        state = download_blobs_from_cloud_storage(
+            bucket_name=release.download_bucket, prefix=prefix, destination_path=release.download_folder
+        )
+        if not state:
+            raise AirflowException("OaWebWorkflow.download failed")
+
+    def make_draft_zenodo_version(self, release: OaWebRelease, **kwargs):
+        """Make a draft Zenodo version of the dataset.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        make_draft_version(release.zenodo, self.conceptrecid)
+
+    def download_twitter_cards(self, release: OaWebRelease, **kwargs):
+        """Download current twitter cards
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        # Download twitter cards
+        blob_name = "twitter.zip"
+        file_path = os.path.join(release.transform_folder, blob_name)
+        download_blob_from_cloud_storage(bucket_name=release.data_bucket_name, blob_name=blob_name, file_path=file_path)
+
+        # Unzip into build
+        unzip_folder_path = os.path.join(release.build_path)
+        with ZipFile(file_path) as zip_file:
+            zip_file.extractall(unzip_folder_path)
+
+    def preprocess_data(self, release: OaWebRelease, **kwargs):
+        """Preprocess data.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        #################
+        # Country
+        #################
+
+        logging.info(f"preprocess_data: country")
+
+        # Load and preprocess data
+        data_path = os.path.join(release.intermediate_path, self.COUNTRY_DATA_FILE)
+        df_data = load_data(data_path)
+        preprocess_data_df("country", df_data)
+
+        # Save to intermediate path
+        data_path = os.path.join(release.intermediate_path, self.COUNTRY_DATA_FILE)
+        records = df_data.to_dict("records")
+        list_to_jsonl_gz(data_path, records)
+
+        #################
+        # Institution
+        #################
+
+        logging.info(f"preprocess_data: institution")
+
+        # Load ROR
+        ror_path = os.path.join(release.download_folder, self.ROR_FILE)
+        ror = load_jsonl(ror_path)
+        tree = ror_to_tree(ror)
+
+        # Load institution data
+        data_path = os.path.join(release.download_folder, self.INSTITUTION_DATA_FILE)
+        df_data = load_data(data_path)
+        preprocess_data_df("institution", df_data)
+        df_data.set_index(["id", "year"], inplace=True, verify_integrity=True)
+        df_data.sort_index(inplace=True)
+
+        df_inst = make_institution_df(ror, start_year=START_YEAR, end_year=END_YEAR)
+        df_inst.update(df_data)
+
+        # Aggregate data
+        aggregate_institutions(tree, df_inst)
+
+        # Save aggregated data
+        data_path = os.path.join(release.intermediate_path, self.INSTITUTION_DATA_FILE)
+        records = df_inst.to_dict("records")
+        list_to_jsonl_gz(data_path, records)
+
+    def build_indexes(self, release: OaWebRelease, **kwargs):
+        """Build unique country and institution indexes.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        for category in self.table_ids:
+            logging.info(f"build_indexes: {category}")
+
+            # Load downloaded index
+            index_name = f"{category}-index.jsonl.gz"
+            index_path = os.path.join(release.download_folder, index_name)
+            df_index = load_data(index_path)
+            preprocess_index_df(category, df_index)
+
+            # Load data file
+            data_path = os.path.join(release.intermediate_path, f"{category}-data.jsonl.gz")
+            df_data = load_data(data_path)
+
+            # Aggregate data file
+            df_index = make_index(category, df_data)
+
+            # Save index to intermediate
+            index_path = os.path.join(release.intermediate_path, index_name)
+            rows: List[Dict] = df_index.to_dict("records")
+            list_to_jsonl_gz(index_path, rows)
+
+    def download_logos(self, release: OaWebRelease, **kwargs):
+        """Download logos and update indexes.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        for category in self.table_ids:
+            logging.info(f"download_logos: {category}")
+
+            # Load index
+            index_path = os.path.join(release.intermediate_path, f"{category}-index.jsonl.gz")
+            df_index = load_data(index_path)
+
+            # Update logos
+            update_index_with_logos(release.build_path, release.assets_path, category, df_index)
+
+            # Save updated index
+            rows: List[Dict] = df_index.to_dict("records")
+            list_to_jsonl_gz(index_path, rows)
+
+    def download_wiki_descriptions(self, release: OaWebRelease, **kwargs):
+        """Download wiki descriptions and update indexes.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        for category in self.table_ids:
+            logging.info(f"download_wiki_descriptions: {category}")
+
+            # Load index
+            index_path = os.path.join(release.intermediate_path, f"{category}-index.jsonl.gz")
+            df_index = load_data(index_path)
+
+            # Update logos
+            update_index_with_wiki_descriptions(df_index)
+
+            # Save updated index
+            rows: List[Dict] = df_index.to_dict("records")
+            list_to_jsonl_gz(index_path, rows)
+
+    def build_datasets(self, release: OaWebRelease, **kwargs):
+        """Transform the queried data into the final format for the open access website.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        # Get versions
+        res = release.zenodo.get_versions(self.conceptrecid, all_versions=1)
+        if res.status_code != 200:
+            raise AirflowException(f"zenodo.get_versions status_code {res.status_code}")
+        versions = res.json()
+
+        # Make required folders
+        entity_index = {category: [] for category in self.table_ids}
+        build_data_path = os.path.join(release.build_path, "data")
+        os.makedirs(build_data_path, exist_ok=True)
+        for category in self.table_ids:
+            logging.info(f"Transforming {category} entity")
+
+            # Load index
+            index_path = os.path.join(release.intermediate_path, f"{category}-index.jsonl.gz")
+            df_index = load_data(index_path)
+
+            data_path = os.path.join(release.intermediate_path, f"{category}-data.jsonl.gz")
+            df_data = load_data(data_path)
+
+            # Make index table
+            entities = make_entities(category, df_index, df_data)
+            entity_index[category] = entities
+
+            # Save index
+            index_path = os.path.join(build_data_path, f"{category}.json")
+            save_index(index_path, entities)
+
+            # Save entities
+            entities_path = os.path.join(build_data_path, category)
+            save_entities(entities_path, entities)
+            logging.info(f"Saved transformed {category} entity")
+
+        # Unwrap lists
+        countries = entity_index["country"]
+        institutions = entity_index["institution"]
+        all_entities = countries + institutions
+
+        # Save all entities as json
+        index_path = os.path.join(build_data_path, f"index.json")
+        save_index(index_path, all_entities)
+
+        # Save COKI Open Access Dataset
+        coki_dataset_path = os.path.join(release.transform_folder, "coki-oa-dataset")
+        save_coki_oa_dataset(coki_dataset_path, countries, institutions)
+
+        # Make stats
+        zenodo_versions = [
+            ZenodoVersion(
+                pendulum.parse(version["created"]),
+                f"https://zenodo.org/record/{version['id']}/files/coki-oa-dataset.zip?download=1",
+            )
+            for version in versions
+        ]
+        last_updated = zenodo_versions[0].release_date.format("D MMMM YYYY")
+        country_stats = make_entity_stats(countries)
+        institution_stats = make_entity_stats(institutions)
+        stats = Stats(START_YEAR, END_YEAR, last_updated, zenodo_versions, country_stats, institution_stats)
+        stats_path = os.path.join(release.build_path, "data")
+        save_stats(stats_path, stats)
+        logging.info(f"Saved stats data")
+
+        # Zip data
+        dst = os.path.join(release.transform_folder, "latest")
+        shutil.copytree(release.build_path, dst)
+        shutil.make_archive(dst, "zip", dst)
+
+    def publish_zenodo_version(self, release: OaWebRelease, **kwargs):
+        """Publish the new Zenodo version of the dataset.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        zenodo = release.zenodo
+        res = zenodo.get_versions(self.conceptrecid, all_versions=0)
+        if res.status_code != 200:
+            raise AirflowException(f"zenodo.get_versions status_code {res.status_code}")
+        draft = res.json()[0]
+        draft_id = draft["id"]
+        if draft["state"] != "unsubmitted":
+            raise AirflowException(f"Latest version is not a draft: {draft_id}")
+
+        file_path = os.path.join(release.transform_folder, "coki-oa-dataset.zip")
+        publish_new_version(release.zenodo, draft_id, file_path)
+
+    def upload_dataset(self, release: OaWebRelease, **kwargs):
+        """Publish the dataset produced by this workflow.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        # upload_file_to_cloud_storage should always rewrite a new version of latest.zip if it exists
+        # object versioning on the bucket will keep the previous versions
+        blob_name = f"{self.version}/latest.zip"
+        file_path = os.path.join(release.transform_folder, "latest.zip")
+        upload_file_to_cloud_storage(
+            bucket_name=release.data_bucket_name, blob_name=blob_name, file_path=file_path, check_blob_hash=False
+        )
+
+    def repository_dispatch(self, release: OaWebRelease, **kwargs):
+        """Trigger a Github repository_dispatch to trigger new website builds.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+
+        token = get_airflow_connection_password(self.GITHUB_TOKEN_CONN)
+        event_types = ["data-update/develop", "data-update/staging", "data-update/production"]
+        for event_type in event_types:
+            trigger_repository_dispatch(token=token, event_type=event_type)
+
+    def cleanup(self, release: OaWebRelease, **kwargs):
+        """Delete all files and folders associated with this release.
+
+        :param release: the release.
+        :param kwargs: the context passed from the PythonOperator. See
+        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
+        passed to this argument.
+        :return: None.
+        """
+        release.cleanup()
+
+
+###############
+# Data classes
+###############
 
 
 @dataclasses.dataclass
@@ -524,6 +1196,11 @@ class Entity:
         return dict_
 
 
+#####################
+# Helper functions
+#####################
+
+
 def save_json(path: str, data: Union[Dict, List]):
     """Save data to JSON.
 
@@ -541,185 +1218,6 @@ def val_empty(val):
         return len(val) == 0
     else:
         return val is None or val == ""
-
-
-def clean_ror_id(ror_id: str):
-    """Remove the https://ror.org/ prefix from a ROR id.
-
-    :param ror_id: original ROR id.
-    :return: cleaned ROR id.
-    """
-
-    return ror_id.replace("https://ror.org/", "")
-
-
-def trigger_repository_dispatch(*, token: str, event_type: str):
-    """Trigger a Github repository dispatch event.
-
-    :param event_type: the event type
-    :param token: the Github token.
-    :return: the response.
-    """
-
-    headers = {
-        "Accept": "application/vnd.github.v3+json",
-        "Authorization": f"token {token}",
-    }
-    data = {"event_type": event_type}
-
-    return requests.post(
-        "https://api.github.com/repos/The-Academic-Observatory/coki-oa-web/dispatches",
-        headers=headers,
-        data=json.dumps(data),
-    )
-
-
-def select_subset(original: Dict, include_keys: Dict):
-    """Select a subset of a dictionary.
-
-    :param original: the original dictionary.
-    :param include_keys: the keys to include.
-    :return:
-    """
-    output = {}
-    for k, v in include_keys.items():
-        if k in original:
-            if isinstance(v, dict):
-                output[k] = select_subset(original[k], v)
-            else:
-                output[k] = original[k]
-
-    return output
-
-
-def get_institution_logo(ror_id: str, url: str, size: str, width: int, fmt: str, build_path) -> Tuple[str, str]:
-    """Get the path to the logo for an institution.
-    If the logo does not exist in the build path yet, download from the Clearbit Logo API tool.
-    If the logo does not exist and failed to download, the path will default to "/unknown.svg".
-
-    :param ror_id: the institution's ROR id
-    :param url: the URL of the company domain + suffix e.g. spotify.com
-    :param size: the image size of the small logo for tables etc.
-    :param width: the width of the image.
-    :param fmt: the image format.
-    :param build_path: the build path for files of this workflow
-    :return: The ROR id and relative path (from build path) to the logo
-    """
-    logo_path = f"/unknown.svg"
-
-    file_path = os.path.join(build_path, "logos", "institution", size, f"{ror_id}.{fmt}")
-    if not os.path.isfile(file_path):
-        clearbit_download_logo(company_url=url, file_path=file_path, size=width, fmt=fmt)
-    if os.path.isfile(file_path):
-        logo_path = make_logo_url(category="institution", entity_id=ror_id, size=size, fmt=fmt)
-
-    return ror_id, logo_path
-
-
-def get_wiki_descriptions(titles: Dict[str, str]) -> List[Tuple[str, str]]:
-    """Get the wikipedia descriptions for the given titles.
-
-    :param titles: Dict with titles as keys and id's (either ror_id or alpha3 country code) as values
-    :return: List with tuples (id, wiki description)
-    """
-    titles_arg = []
-    for title, entity_id in titles.items():
-        # URL encode title if it is not encoded yet
-        if title == urllib.parse.unquote(title):
-            titles_arg.append(urllib.parse.quote(title))
-        # Append title directly if it is already encoded and not empty
-        else:
-            titles_arg.append(title)
-
-    # Confirm that there is a max of 20 titles, the limit for the wikipedia API
-    assert len(titles_arg) <= 20
-
-    # Extract descriptions using the Wikipedia API
-    url = f"https://en.wikipedia.org/w/api.php?action=query&format=json&prop=extracts&titles={'%7C'.join(titles_arg)}&redirects=1&exintro=1&explaintext=1"
-    response = requests.get(url)
-    if response.status_code != 200:
-        raise AirflowException(f"Unsuccessful retrieving wikipedia extracts, url: {url}")
-    response_json = response.json()
-    pages = response_json["query"]["pages"]
-
-    # Create mapping between redirected/normalized page title and original page title
-    redirects = {}
-    for title in response_json["query"].get("redirects", []):
-        redirects[title["to"]] = title["from"]
-    normalized = {}
-    for title in response_json["query"].get("normalized", []):
-        normalized[title["to"]] = title["from"]
-
-    # Resolve redirects referring to each other to 1 redirect
-    for key, value in redirects.copy().items():
-        if value in redirects:
-            redirects[key] = redirects.pop(value)
-
-    # Create mapping between entity_id and decoded page title.
-    decoded_titles = {urllib.parse.unquote(k): v for k, v in titles.items()}
-    descriptions = []
-    for page_id, page in pages.items():
-        page_title = page["title"]
-
-        # Get page_title from redirected/normalized if it is present
-        page_title = redirects.get(page_title, page_title)
-        page_title = normalized.get(page_title, page_title)
-
-        # Link original title to description
-        entity_id = decoded_titles[urllib.parse.unquote(page_title)]
-
-        # Get description and clean up
-        description = page.get("extract", "")
-        if description:
-            description = remove_text_between_brackets(description)
-            description = shorten_text_full_sentences(description)
-
-        descriptions.append((entity_id, description))
-    return descriptions
-
-
-def remove_text_between_brackets(text: str) -> str:
-    """Remove any text between (nested) brackets.
-    If there is a space after the opening bracket, this is removed as well.
-    E.g. 'Like this (foo, (bar)) example' -> 'Like this example'
-
-    :param text: The text to modify
-    :return: The modified text
-    """
-    new_text = []
-    nested = 0
-    for char in text:
-        if char == "(":
-            nested += 1
-            new_text = new_text[:-1] if new_text and new_text[-1] == " " else new_text
-        elif (char == ")") and nested:
-            nested -= 1
-        elif nested == 0:
-            new_text.append(char)
-    return "".join(new_text).strip()
-
-
-def shorten_text_full_sentences(text: str, *, char_limit: int = 300) -> str:
-    """Shorten a text to as many complete sentences as possible, while the total number of characters stays below
-    the char_limit.
-    Always return at least one sentence, even if this exceeds the char_limit.
-
-    :param text: A string with the complete text
-    :param char_limit: The max number of characters
-    :return: The shortened text.
-    """
-    # Create list of sentences
-    sentences = nltk.tokenize.sent_tokenize(text)
-
-    # Add sentences until char limit is reached
-    sentences_output = []
-    total_len = 0
-    for sentence in sentences:
-        total_len += len(sentence)
-        if (total_len > char_limit) and sentences_output:
-            break
-        sentences_output.append(sentence)
-    return " ".join(sentences_output)
 
 
 def bq_query_to_gcs(*, query: str, project_id: str, destination_uri: str, location: str = "us") -> bool:
@@ -740,25 +1238,15 @@ def bq_query_to_gcs(*, query: str, project_id: str, destination_uri: str, locati
 
     # Create and run extraction job
     source_table_id = f"{project_id}.{query_job.destination.dataset_id}.{query_job.destination.table_id}"
-    extract_job_config = bigquery.ExtractJobConfig()
-    extract_job_config.destination_format = bigquery.DestinationFormat.NEWLINE_DELIMITED_JSON
+    config = bigquery.ExtractJobConfig()
+    config.destination_format = bigquery.DestinationFormat.NEWLINE_DELIMITED_JSON
+    config.compression = bigquery.Compression.GZIP
     extract_job: bigquery.ExtractJob = client.extract_table(
-        source_table_id, destination_uri, job_config=extract_job_config, location=location
+        source_table_id, destination_uri, job_config=config, location=location
     )
     extract_job.result()
 
     return query_job.state == "DONE" and extract_job.state == "DONE"
-
-
-def clean_url(url: str) -> str:
-    """Remove path and query from URL.
-
-    :param url: the url.
-    :return: the cleaned url.
-    """
-
-    p = urlparse(url)
-    return f"{p.scheme}://{p.netloc}/"
 
 
 def save_as_jsonl(output_path: str, iterable: List[Dict]):
@@ -774,244 +1262,19 @@ def save_as_jsonl(output_path: str, iterable: List[Dict]):
             writer.write_all(iterable)
 
 
-def make_logo_url(*, category: str, entity_id: str, size: str, fmt: str) -> str:
-    """Make a logo url.
+#######################################
+# Institution Hierarchical Aggregation
+#######################################
 
-    :param category: the entity category: country or institution.
-    :param entity_id: the entity id.
-    :param size: the size of the logo: s or l.
-    :param fmt: the format of the logo.
-    :return: the logo url.
+
+def clean_ror_id(ror_id: str):
+    """Remove the https://ror.org/ prefix from a ROR id.
+
+    :param ror_id: original ROR id.
+    :return: cleaned ROR id.
     """
 
-    return f"/logos/{category}/{size}/{entity_id}.{fmt}"
-
-
-class Zenodo:
-    def __init__(self, host: str = "https://zenodo.org", access_token: str = None, timeout: int = 60):
-        self.host = host
-        self.access_token = access_token
-        self.timeout = timeout
-
-    def make_url(self, path: str):
-        # Remove last / from host
-        host = self.host
-        if self.host[-1] == "/":
-            host = self.host[:-1]
-
-        # Add leading / to path
-        if path[0] != "/":
-            path = f"/{path}"
-
-        return f"{host}{path}"
-
-    def get_versions(self, conceptrecid: int, all_versions: int = 0, size: int = 10, sort: str = "mostrecent"):
-        query = f"conceptrecid:{conceptrecid}"
-        return requests.get(
-            self.make_url("/api/deposit/depositions"),
-            params={
-                "q": query,
-                "all_versions": all_versions,
-                "access_token": self.access_token,
-                "sort": sort,
-                "size": size,
-            },
-            timeout=self.timeout,
-        )
-
-    def create_new_version(self, id: int):
-        url = self.make_url(f"/api/deposit/depositions/{id}/actions/newversion")
-        return requests.post(url, params={"access_token": self.access_token})
-
-    def get_deposition(self, id: int):
-        url = self.make_url(f"/api/deposit/depositions/{id}")
-        return requests.get(url, params={"access_token": self.access_token})
-
-    def delete_file(self, id: int, file_id: str):
-        url = self.make_url(f"/api/deposit/depositions/{id}/files/{file_id}")
-        return requests.delete(url, params={"access_token": self.access_token})
-
-    def update(self, id: int, data: Dict):
-        url = self.make_url(f"/api/deposit/depositions/{id}")
-        headers = {"Content-Type": "application/json"}
-        return requests.put(url, data=json.dumps(data), headers=headers, params={"access_token": self.access_token})
-
-    def upload_file(self, id: int, file_path: str):
-        url = self.make_url(f"/api/deposit/depositions/{id}/files")
-        data = {"name": os.path.basename(file_path)}
-        files = {"file": open(file_path, "rb")}
-        return requests.post(url, data=data, files=files, params={"access_token": self.access_token})
-
-    def publish(self, id: int):
-        url = self.make_url(f"/api/deposit/depositions/{id}/actions/publish")
-        return requests.post(url, params={"access_token": self.access_token})
-
-
-def make_draft_version(zenodo: Zenodo, conceptrecid: int):
-    # Make new draft version
-    res = zenodo.get_versions(conceptrecid)
-    if res.status_code != 200:
-        raise AirflowException(f"zenodo.get_versions status_code {res.status_code}")
-
-    versions = res.json()
-    if len(versions) == 0:
-        raise AirflowException(f"make_draft_version: at least 1 version must exist")
-
-    latest = versions[0]
-    draft_id = latest["id"]
-    state = latest["state"]
-    if state == "done":
-        # If published then create a new draft
-        res = zenodo.create_new_version(draft_id)
-        if res.status_code != 201:
-            raise AirflowException(f"zenodo.create_new_version status_code {res.status_code}")
-        draft_id = int(res.json()["links"]["latest_draft"].split("/")[-1])
-
-    # Fetch draft deposition
-    res = zenodo.get_deposition(draft_id)
-    if res.status_code != 200:
-        raise AirflowException(f"zenodo.get_deposition status_code {res.status_code}")
-
-    # Update metadata
-    draft = res.json()
-    publication_date = pendulum.now().format("YYYY-MM-DD")
-    metadata = draft["metadata"]
-    metadata["publication_date"] = publication_date
-    metadata["version"] = publication_date
-    data = {"metadata": metadata}
-    res = zenodo.update(draft_id, data)
-    if res.status_code != 200:
-        raise AirflowException(f"zenodo.update status_code {res.status_code}")
-
-
-def publish_new_version(zenodo: Zenodo, draft_id: int, file_path: str):
-    # Get full deposition which contains files
-    res = zenodo.get_deposition(draft_id)
-    if res.status_code != 200:
-        raise AirflowException(f"zenodo.get_deposition {res.status_code}")
-    draft = res.json()
-
-    # Delete existing files
-    for file in draft["files"]:
-        file_id = file["id"]
-        res = zenodo.delete_file(draft_id, file_id)
-        if res.status_code != 204:
-            raise AirflowException(f"zenodo.delete_file status_code {res.status_code}")
-
-    # Upload new file
-    res = zenodo.upload_file(draft_id, file_path)
-    if res.status_code != 201:
-        raise AirflowException(f"zenodo.upload_file status_code {res.status_code}")
-
-    # Publish
-    res = zenodo.publish(draft_id)
-    if res.status_code != 202:
-        raise AirflowException(f"zenodo.publish status_code {res.status_code}")
-
-
-def load_data(file_path: str) -> pd.DataFrame:
-    """Load a country or institution data file into a Pandas DataFrame.
-
-    :param file_path: the path to the file to load.
-    :return: the Pandas Dataframe.
-    """
-
-    data = load_jsonl(file_path)
-    return pd.DataFrame(data)
-
-
-#
-# def preprocess_df(category: str, df: pd.DataFrame):
-#     """Pre-process the data frame.
-#
-#     :param category: the category.
-#     :param df: the dataframe.
-#     :return: the Pandas Dataframe.
-#     """
-#
-#     # Convert data types
-#     df["date"] = pd.to_datetime(df["date"])
-#     df.fillna("", inplace=True)
-#     for column in df.columns:
-#         if column.startswith("n_"):
-#             df[column] = pd.to_numeric(df[column])
-#
-#     # Clean RoR ids
-#     if category == "institution":
-#         # Remove columns not used for institutions
-#         df.drop(columns=["alpha2"], inplace=True, errors="ignore")
-#
-#         # Clean RoR ids
-#         df["id"] = df["id"].apply(lambda i: clean_ror_id(i))
-#
-#         # Parse identifiers
-#         preferred_key = "preferred"
-#         identifiers = []
-#         for i, row in df.iterrows():
-#             # Parse identifier for each entry
-#             ent_ids = []
-#             ids_dict = row["identifiers"]
-#
-#             # Add ROR id
-#             ror_id = row["id"]
-#             ent_ids.append({"id": ror_id, "type": "ROR", "url": f"https://ror.org/{ror_id}"})
-#
-#             # Parse other ids
-#             for k, v in ids_dict.items():
-#                 url = None
-#                 id_type = k
-#                 if id_type != "OrgRef":
-#                     if preferred_key in v:
-#                         id_value = v[preferred_key]
-#                     else:
-#                         id_value = v["all"][0]
-#
-#                     # Create URLs
-#                     if id_type == "ISNI":
-#                         url = f"https://isni.org/isni/{id_value}"
-#                     elif id_type == "Wikidata":
-#                         url = f"https://www.wikidata.org/wiki/{id_value}"
-#                     elif id_type == "GRID":
-#                         url = f"https://grid.ac/institutes/{id_value}"
-#                     elif id_type == "FundRef":
-#                         url = f"https://api.crossref.org/funders/{id_value}"
-#
-#                     ent_ids.append({"id": id_value, "type": id_type, "url": url})
-#             identifiers.append(ent_ids)
-#         df["identifiers"] = identifiers
-#
-#     if category == "country":
-#         # Remove columns not used for countries
-#         df.drop(columns=["url", "institution_types", "country", "identifiers"], inplace=True, errors="ignore")
-
-
-def preprocess_df(category: str, df: pd.DataFrame):
-    """Pre-process the data frame.
-
-    :param category: the category.
-    :param df: the dataframe.
-    :return: the Pandas Dataframe.
-    """
-
-    # Convert data types
-    # df["date"] = pd.to_datetime(df["date"])
-    df["year"] = pd.to_numeric(df["year"])
-    df.fillna("", inplace=True)
-    for column in df.columns:
-        if column.startswith("n_"):
-            df[column] = pd.to_numeric(df[column])
-
-    # Clean RoR ids
-    if category == "institution":
-        # Remove columns not used for institutions
-        # df.drop(columns=["alpha2"], inplace=True, errors="ignore")
-
-        # Clean RoR ids
-        df["id"] = df["id"].apply(lambda i: clean_ror_id(i))
-
-    # df.drop(columns=["repositories"], inplace=True, errors="ignore")
-    df.set_index(["id", "year"], inplace=True, verify_integrity=True)
-    df.sort_index(inplace=True)
+    return ror_id.replace("https://ror.org/", "")
 
 
 def ror_to_tree(ror: List[Dict]) -> treelib.Tree:
@@ -1056,6 +1319,42 @@ def ror_to_tree(ror: List[Dict]) -> treelib.Tree:
     return tree
 
 
+def aggregate_institutions(tree: treelib.Tree, df: pd.DataFrame) -> None:
+    """Aggregate institutional data in a hierarchical fashion. Data is aggregated in place.
+
+    :param tree: the ROR tree.
+    :param df: the data to aggregate in a Pandas DataFrame.
+    :return: None.
+    """
+
+    seen = set()
+    root_id = "root"
+    leaves = tree.leaves(root_id)
+    n_nodes = len(tree.nodes)
+
+    for leaf in leaves:
+        # If leaf already processed then continue
+        leaf_id = leaf.identifier
+        if leaf_id in seen:
+            continue
+
+        # Mark that leaf has been seen
+        seen.add(leaf_id)
+
+        # If leaf of root, no aggregation needed
+        p = tree.parent(leaf_id)
+        if p.identifier == root_id:
+            continue
+
+        # Get aggregated stats for leaf
+        aggregate_bottom_up(leaf_id, tree, df, seen)
+
+        # Print progress
+        n_processed = len(seen) + 1
+        p_processed = n_processed / n_nodes * 100
+        print(f"Progress \r{n_processed} / {n_nodes}: {p_processed:.2f}%")
+
+
 def aggregate_bottom_up(node_id: str, tree: treelib.Tree, df: pd.DataFrame, seen: Set) -> None:
     """Aggregate from the leaves to the root of a subtree, starting at the parent of node_id.
 
@@ -1088,16 +1387,27 @@ def aggregate_bottom_up(node_id: str, tree: treelib.Tree, df: pd.DataFrame, seen
     aggregate_bottom_up(parent.identifier, tree, df, seen)
 
 
-def repositories_sort_key(col):
-    if is_string_dtype(col.dtype):
-        return col.str.lower()
-    return col
+def traverse_children(parent_id: str, tree: treelib.Tree, df: pd.DataFrame, seen: Set) -> None:
+    """Recursively traverse the children of parent_id aggregating the parent and child nodes from the bottom of the
+    tree, to the top.
 
+    :param parent_id: the parent node id.
+    :param tree: the ROR tree.
+    :param df: the Pandas Dataframe with the data.
+    :param seen: a list of node ids that have been seen.
+    :return: None.
+    """
 
-def repositories_merge_category(cat):
-    if cat in {"Aggregator", "Unknown"}:
-        return "Other Internet"
-    return cat
+    # Get children
+    children = []
+    for child in tree.children(parent_id):
+        traverse_children(child.identifier, tree, df, seen)
+        children.append(child.identifier)
+        seen.add(child.identifier)
+
+    # Aggregate parent and children
+    if len(children):
+        aggregate_parent_and_children(parent_id, children, df)
 
 
 def aggregate_parent_and_children(pid: str, child_ids: List, df: pd.DataFrame) -> None:
@@ -1185,63 +1495,113 @@ def aggregate_repositories(ror_id: str, df: pd.DataFrame) -> pd.DataFrame:
     return df
 
 
-def traverse_children(parent_id: str, tree: treelib.Tree, df: pd.DataFrame, seen: Set) -> None:
-    """Recursively traverse the children of parent_id aggregating the parent and child nodes from the bottom of the
-    tree, to the top.
+def repositories_sort_key(col):
+    if is_string_dtype(col.dtype):
+        return col.str.lower()
+    return col
 
-    :param parent_id: the parent node id.
-    :param tree: the ROR tree.
-    :param df: the Pandas Dataframe with the data.
-    :param seen: a list of node ids that have been seen.
-    :return: None.
+
+def repositories_merge_category(cat):
+    if cat in {"Aggregator", "Unknown"}:
+        return "Other Internet"
+    return cat
+
+
+######################
+# Transform data
+######################
+
+
+def load_data(file_path: str) -> pd.DataFrame:
+    """Load a country or institution data file into a Pandas DataFrame.
+
+    :param file_path: the path to the file to load.
+    :return: the Pandas Dataframe.
     """
 
-    # Get children
-    children = []
-    for child in tree.children(parent_id):
-        traverse_children(child.identifier, tree, df, seen)
-        children.append(child.identifier)
-        seen.add(child.identifier)
-
-    # Aggregate parent and children
-    if len(children):
-        aggregate_parent_and_children(parent_id, children, df)
+    data = load_jsonl(file_path)
+    return pd.DataFrame(data)
 
 
-def aggregate_institutions(tree: treelib.Tree, df: pd.DataFrame) -> None:
-    """Aggregate institutional data in a hierarchical fashion. Data is aggregated in place.
+def preprocess_data_df(category: str, df: pd.DataFrame):
+    """Pre-process the data frame.
 
-    :param tree: the ROR tree.
-    :param df: the data to aggregate in a Pandas DataFrame.
-    :return: None.
+    :param category: the category.
+    :param df: the dataframe.
+    :return: the Pandas Dataframe.
     """
 
-    seen = set()
-    root_id = "root"
-    leaves = tree.leaves(root_id)
-    n_nodes = len(tree.nodes)
+    # Convert data types
+    df["year"] = pd.to_numeric(df["year"])
+    df.fillna("", inplace=True)
+    for column in df.columns:
+        if column.startswith("n_"):
+            df[column] = pd.to_numeric(df[column])
 
-    for leaf in leaves:
-        # If leaf already processed then continue
-        leaf_id = leaf.identifier
-        if leaf_id in seen:
-            continue
+    # Category specific processing
+    if category == "institution":
+        # Clean RoR ids
+        df["id"] = df["id"].apply(lambda i: clean_ror_id(i))
 
-        # Mark that leaf has been seen
-        seen.add(leaf_id)
+    # Make multi index based on id and year
+    df.set_index(["id", "year"], inplace=True, verify_integrity=True)
+    df.sort_index(inplace=True)
 
-        # If leaf of root, no aggregation needed
-        p = tree.parent(leaf_id)
-        if p.identifier == root_id:
-            continue
 
-        # Get aggregated stats for leaf
-        aggregate_bottom_up(leaf_id, tree, df, seen)
+def preprocess_index_df(category: str, df: pd.DataFrame):
+    """Pre-process the index data frame.
 
-        # Print progress
-        n_processed = len(seen) + 1
-        p_processed = n_processed / n_nodes * 100
-        print(f"Progress \r{n_processed} / {n_nodes}: {p_processed:.2f}%")
+    :param category: the category.
+    :param df: the dataframe.
+    :return: the Pandas Dataframe.
+    """
+
+    # Convert data types
+    df.fillna("", inplace=True)
+
+    # Clean RoR ids
+    if category == "institution":
+        # Remove columns not used for institutions
+        df.drop(columns=["alpha2"], inplace=True, errors="ignore")
+
+        # Clean RoR ids
+        df["id"] = df["id"].apply(lambda i: clean_ror_id(i))
+
+        # Parse identifiers
+        preferred_key = "preferred"
+        identifiers = []
+        for i, row in df.iterrows():
+            # Parse identifier for each entry
+            ent_ids = []
+            ids_dict = row["identifiers"]
+
+            # Add ROR id
+            ror_id = row["id"]
+            ent_ids.append({"id": ror_id, "type": "ROR", "url": f"https://ror.org/{ror_id}"})
+
+            # Parse other ids
+            for k, v in ids_dict.items():
+                url = None
+                id_type = k
+                if id_type != "OrgRef":
+                    if preferred_key in v:
+                        id_value = v[preferred_key]
+                    else:
+                        id_value = v["all"][0]
+
+                    # Create URLs
+                    if id_type == "ISNI":
+                        url = f"https://isni.org/isni/{id_value}"
+                    elif id_type == "Wikidata":
+                        url = f"https://www.wikidata.org/wiki/{id_value}"
+                    elif id_type == "GRID":
+                        url = f"https://grid.ac/institutes/{id_value}"
+                    elif id_type == "FundRef":
+                        url = f"https://api.crossref.org/funders/{id_value}"
+
+                    ent_ids.append({"id": id_value, "type": id_type, "url": url})
+            identifiers.append(ent_ids)
+        df["identifiers"] = identifiers
 
 
 def make_institution_df(
@@ -1295,30 +1655,198 @@ def make_institution_df(
     return df
 
 
-class OaWebRelease(SnapshotRelease):
-    def __init__(
-        self,
-        *,
-        dag_id: str,
-        release_date: pendulum.DateTime,
-        data_bucket_name: str,
-        zenodo: Zenodo = Zenodo(),
-    ):
-        """Create an OaWebRelease instance.
+def make_index(category: str, df_data: pd.DataFrame):
+    """Make the data for the index tables.
 
-        :param dag_id: the dag id.
-        :param release_date: the release date.
-        :param zenodo: the zenodo instance.
-        """
+    :param category: the category, i.e. country or institution.
+    :param df_data: Pandas dataframe with all data points.
+    :return:
+    """
 
-        super().__init__(dag_id=dag_id, release_date=release_date)
-        self.zenodo = zenodo
-        self.data_bucket_name = data_bucket_name
-        self.assets_path = module_file_path("academic_observatory_workflows.workflows.data.oa_web_workflow")
+    # Create aggregate
+    agg = {}
+    for column in df_data.columns:
+        if column.startswith("n_"):
+            agg[column] = "sum"
+        else:
+            agg[column] = "first"
 
-    @property
-    def build_path(self):
-        return os.path.join(self.transform_folder, "build")
+    # Create aggregate
+    df_index_table = (
+        df_data.reset_index()
+        .groupby(["id"])
+        .agg(
+            agg,
+            index=False,
+        )
+    )
+
+    # Exclude countries with small samples
+    df_index_table = df_index_table[df_index_table["n_outputs"] >= INCLUSION_THRESHOLD[category]]
+
+    # Add percentages to dataframe
+    update_df_with_percentages(df_index_table, PERCENTAGE_FIELD_KEYS)
+
+    # Sort from highest oa percentage to lowest
+    df_index_table.sort_values(by=["n_outputs_open"], ascending=False, inplace=True)
+
+    # Add category
+    df_index_table["category"] = category
+
+    # Remove date, year and repositories
+    df_index_table.drop(columns=["year", "repositories"], inplace=True)
+
+    return df_index_table
+
+
+def update_df_with_percentages(df: pd.DataFrame, keys: List[Tuple[str, str]]):
+    """Calculate percentages for fields in a Pandas dataframe.
+
+    :param df: the Pandas dataframe.
+    :param keys: they keys to calculate percentages for.
+    :return: None.
+    """
+
+    for numerator_key, denominator_key in keys:
+        p_key = f"p_{numerator_key}"
+        df[p_key] = round(df[f"n_{numerator_key}"] / df[denominator_key] * 100, 2)
+
+        # Fill in NaN caused by denominator of zero
+        df[p_key] = df[p_key].fillna(0)
+
+
+def select_subset(original: Dict, include_keys: Dict):
+    """Select a subset of a dictionary.
+
+    :param original: the original dictionary.
+    :param include_keys: the keys to include.
+    :return:
+    """
+    output = {}
+    for k, v in include_keys.items():
+        if k in original:
+            if isinstance(v, dict):
+                output[k] = select_subset(original[k], v)
+            else:
+                output[k] = original[k]
+
+    return output
+
+
+def save_index(file_path: str, entities: List[Entity]):
+    """Save an index file.
+
+    :param file_path: the file path where the index should be saved.
+    :param entities: a list of entities.
+    :return: None.
+    """
+
+    # Convert entities to dictionaries and select a subset of fields
+    subset = {
+        "id": None,
+        "name": None,
+        "logo_s": None,
+        "category": None,
+        "country": None,
+        "subregion": None,
+        "region": None,
+        "institution_types": None,
+        "acronyms": None,
+        "stats": {
+            "n_outputs": None,
+            "n_outputs_open": None,
+            "p_outputs_open": None,
+            "p_outputs_publisher_open_only": None,
+            "p_outputs_both": None,
+            "p_outputs_other_platform_open_only": None,
+            "p_outputs_closed": None,
+        },
+    }
+    data = [select_subset(entity.to_dict(), subset) for entity in entities]
+
+    # Save as JSON
+    save_json(file_path, data)
+
+
+def make_entities(category: str, df_index: pd.DataFrame, df_data: pd.DataFrame) -> List[Entity]:
+    """Make entities.
+
+    :param category: the entity category.
+    :param df_index: the index dataframe.
+    :param df_data: the data dataframe.
+    :return: the Entity objects.
+    """
+
+    entities = []
+    key_id = "id"
+    key_year = "year"
+    key_date = "date"
+    key_records = "records"
+    ts_groups = df_data.groupby([key_id])
+    for entity_id, df_group in ts_groups:
+        # Exclude countries and institutions with small num outputs
+        total_outputs = df_group["n_outputs"].sum()
+        if total_outputs >= INCLUSION_THRESHOLD[category]:
+            update_df_with_percentages(df_group, PERCENTAGE_FIELD_KEYS)
+            df_group = df_group.sort_values(by=[key_year])
+            df_group = df_group.loc[:, ~df_group.columns.str.contains("^Unnamed")]
+
+            # Create entity
+            entity_dict: Dict = df_index.loc[df_index[key_id] == entity_id].to_dict(key_records)[0]
+            entity = Entity.from_dict(entity_dict)
+            entity.stats = PublicationStats.from_dict(entity_dict)
+
+            # Make timeseries data
+            years = []
+            rows: List[Dict] = df_group.to_dict(key_records)
+            for row in rows:
+                year = int(row.get(key_year))
+                date = row.get(key_date)
+                stats = PublicationStats.from_dict(row)
+                years.append(Year(year=year, date=date, stats=stats))
+            entity.years = years
+
+            # Make repositories
+            repositories = []
+            for row in rows:
+                repositories += row["repositories"]
+            df_repos = pd.DataFrame(repositories, columns=["id", "total_outputs", "category", "home_repo"])
+            df_repos["total_outputs"] = pd.to_numeric(df_repos["total_outputs"])
+            df_repos["category"] = df_repos["category"].apply(repositories_merge_category)
+            df_repos = (
+                df_repos.groupby(["id"], as_index=False)
+                .agg(
+                    {
+                        "id": "first",
+                        "total_outputs": "sum",
+                        "category": "first",
+                        "home_repo": "first",
+                    }
+                )
+                .sort_values(
+                    by=["total_outputs", "id"], ascending=[False, True], inplace=False, key=repositories_sort_key
+                )
+            )
+            repositories = df_repos.to_dict(key_records)
+            repositories = [Repository.from_dict(repo) for repo in repositories]
+
+            # Select a maximum number of repositories, however, make sure that all repositories that belong
+            # to the given institution are always present
+            entity.repositories = []
+            for repo in repositories:
+                if len(entity.repositories) <= MAX_REPOSITORIES or repo.home_repo:
+                    entity.repositories.append(repo)
+
+            # Set min and max years for data
+            entity.start_year = years[0].year
+            entity.end_year = years[-1].year
+
+            entities.append(entity)
+
+    # Ensure that entities are sorted based on p_outputs_open
+    entities = sorted(entities, key=lambda e: e.stats.p_outputs_open, reverse=True)
+
+    return entities
 
 
 def make_entity_stats(entities: List[Entity]) -> EntityStats:
@@ -1366,902 +1894,267 @@ def make_entity_stats(entities: List[Entity]) -> EntityStats:
     )
 
 
-class OaWebWorkflow(Workflow):
-    PERCENTAGE_FIELD_KEYS = [
-        ("outputs_open", "n_outputs"),
-        ("outputs_both", "n_outputs"),
-        ("outputs_closed", "n_outputs"),
-        ("outputs_publisher_open", "n_outputs"),
-        ("outputs_publisher_open_only", "n_outputs"),
-        ("outputs_other_platform_open", "n_outputs"),
-        ("outputs_other_platform_open_only", "n_outputs"),
-        ("outputs_oa_journal", "n_outputs_publisher_open"),
-        ("outputs_hybrid", "n_outputs_publisher_open"),
-        ("outputs_no_guarantees", "n_outputs_publisher_open"),
-        ("outputs_preprint", "n_outputs_other_platform_open"),
-        ("outputs_domain", "n_outputs_other_platform_open"),
-        ("outputs_institution", "n_outputs_other_platform_open"),
-        ("outputs_public", "n_outputs_other_platform_open"),
-        ("outputs_other_internet", "n_outputs_other_platform_open"),
-    ]
-    DATA_BUCKET = "oa_web_data_bucket"
-    GITHUB_TOKEN_CONN = "oa_web_github_token"
-    ZENODO_TOKEN_CONN = "oa_web_zenodo_token"
+#################
+# Logo fetching
+#################
 
-    """The OaWebWorkflow generates data files for the COKI Open Access Dashboard.
 
-    The figure below illustrates the generated data and notes about what each file is used for.
-    .
-    ├── data: data
-    │   ├── index.json: used by the Cloudflare Worker search and filtering API.
-    │   ├── autocomplete.json: used for the website search functionality. Copied into public/data folder.
-    │   ├── country: individual entity statistics files for countries. Used to build each country page.
-    │   │   ├── ALB.json
-    │   │   ├── ARE.json
-    │   │   └── ARG.json
-    │   ├── country.json: used to create the country table. First 18 countries used to build first page of country table
-    │   │                 and then this file is included in the public folder and downloaded by the client to enable the
-    │   │                 other pages of the table to be displayed. Copied into public/data folder.
-    │   ├── institution: individual entity statistics files for institutions. Used to build each institution page.
-    │   │   ├── 05ykr0121.json
-    │   │   ├── 05ym42410.json
-    │   │   └── 05ynxx418.json
-    │   ├── institution.json: used to create the institution table. First 18 institutions used to build first page of institution table
-    │   │                     and then this file is included in the public folder and downloaded by the client to enable the
-    │   │                     other pages of the table to be displayed. Copied into public/data folder.
-    │   └── stats.json: global statistics, e.g. the minimum and maximum date for the dataset, when it was last updated etc.
-    └── logos: country and institution logos. Copied into public/logos folder.
-        ├── country
-        │   ├── l: large logos displayed on country pages.
-        │   │   ├── ALB.svg
-        │   │   ├── ARE.svg
-        │   │   └── ARG.svg
-        │   └── s: small logos displayed in country table.
-        │       ├── ALB.svg
-        │       ├── ARE.svg
-        │       └── ARG.svg
-        └── institution
-            ├── l: large logos displayed on institution pages.
-            │   ├── 05ykr0121.jpg
-            │   ├── 05ym42410.jpg
-            │   └── 05ynxx418.jpg
-            └── s: small logos displayed in institution table.
-                ├── 05ykr0121.jpg
-                ├── 05ym42410.jpg
-                └── 05ynxx418.jpg
+def make_logo_url(*, category: str, entity_id: str, size: str, fmt: str) -> str:
+    """Make a logo url.
+
+    :param category: the entity category: country or institution.
+    :param entity_id: the entity id.
+    :param size: the size of the logo: s or l.
+    :param fmt: the format of the logo.
+    :return: the logo url.
     """
 
-    # Set the number of titles for which wiki descriptions are retrieved at once, the API can return max 20 extracts.
-    WIKI_MAX_TITLES = 20
-
-    def __init__(
-        self,
-        *,
-        input_project_id: str = "academic-observatory",
-        output_project_id: str = "academic-observatory",
-        dag_id: str = "oa_web_workflow",
-        start_date: Optional[pendulum.DateTime] = pendulum.datetime(2021, 5, 2),
-        schedule_interval: Optional[str] = "@weekly",
-        catchup: Optional[bool] = False,
-        ext_dag_id: str = "doi",
-        table_ids: List[str] = None,
-        airflow_vars: List[str] = None,
-        airflow_conns: List[str] = None,
-        agg_dataset_id: str = "observatory",
-        ror_dataset_id: str = "ror",
-        settings_dataset_id: str = "settings",
-        version: str = "v6",
-        conceptrecid: int = 6399462,
-        zenodo_host: str = "https://zenodo.org",
-    ):
-        """Create the OaWebWorkflow.
-
-        :param project_id: the Google Cloud project id.
-        :param dag_id: the DAG id.
-        :param start_date: the start date.
-        :param schedule_interval: the schedule interval.
-        :param catchup: whether to catchup or not.
-        :param ext_dag_id: the DAG id to wait for.
-        :param table_ids: the table ids.
-        :param airflow_vars: required Airflow Variables.
-        :param airflow_conns: required Airflow Connections.
-        :param agg_dataset_id: the id of the dataset where the Academic Observatory aggregated data lives.
-        :param ror_dataset_id: the id of the dataset containing the ROR table.
-        :param settings_dataset_id: the id of the settings dataset, which contains the country table.
-        :param version: the dataset version published by this workflow. The Github Action pulls from a specific dataset
-        version: https://github.com/The-Academic-Observatory/coki-oa-web/blob/develop/.github/workflows/build-on-data-update.yml#L68-L74.
-        This is so that when breaking changes are made to the schema, the web application won't break.
-        :param conceptrecid: the Zenodo Concept Record ID for the COKI Open Access Dataset. The Concept Record ID is
-        the last set of numbers from the Concept DOI.
-        :param zenodo_host: the Zenodo hostname, can be changed to https://sandbox.zenodo.org for testing.
-        """
-
-        if airflow_vars is None:
-            airflow_vars = [
-                AirflowVars.DATA_PATH,
-                AirflowVars.PROJECT_ID,
-                AirflowVars.DATA_LOCATION,
-                AirflowVars.DOWNLOAD_BUCKET,
-                AirflowVars.TRANSFORM_BUCKET,
-                self.DATA_BUCKET,
-            ]
-
-        if airflow_conns is None:
-            airflow_conns = [self.GITHUB_TOKEN_CONN, self.ZENODO_TOKEN_CONN]
-
-        super().__init__(
-            dag_id=dag_id,
-            start_date=start_date,
-            schedule_interval=schedule_interval,
-            catchup=catchup,
-            airflow_vars=airflow_vars,
-            airflow_conns=airflow_conns,
-            tags=[Tag.academic_observatory],
-        )
-        self.input_project_id = input_project_id
-        self.output_project_id = output_project_id
-        self.agg_dataset_id = agg_dataset_id
-        self.ror_dataset_id = ror_dataset_id
-        self.settings_dataset_id = settings_dataset_id
-        self.table_ids = table_ids
-        self.version = version
-        self.conceptrecid = conceptrecid
-        self.zenodo_host = zenodo_host
-        if table_ids is None:
-            self.table_ids = ["country", "institution"]
-
-        self.add_operator(
-            ExternalTaskSensor(task_id=f"{ext_dag_id}_sensor", external_dag_id=ext_dag_id, mode="reschedule")
-        )
-        self.add_setup_task(self.check_dependencies)
-        self.add_task(self.query)
-        self.add_task(self.download)
-        self.add_task(self.make_draft_zenodo_version)
-        self.add_task(self.download_twitter_cards)
-        self.add_task(self.transform)
-        self.add_task(self.publish_zenodo_version)
-        self.add_task(self.upload_dataset)
-        self.add_task(self.repository_dispatch)
-        self.add_task(self.cleanup)
-
-    ######################################
-    # Airflow tasks
-    ######################################
-
-    def make_release(self, **kwargs) -> OaWebRelease:
-        """Make release instances. The release is passed as an argument to the function (TelescopeFunction) that is
-        called in 'task_callable'.
-
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: A list of OaWebRelease instances
-        """
-
-        release_date = make_release_date(**kwargs)
-        data_bucket_name = Variable.get(self.DATA_BUCKET)
-        zenodo_token = get_airflow_connection_password(self.ZENODO_TOKEN_CONN)
-        zenodo = Zenodo(host=self.zenodo_host, access_token=zenodo_token)
-
-        return OaWebRelease(
-            dag_id=self.dag_id,
-            data_bucket_name=data_bucket_name,
-            release_date=release_date,
-            zenodo=zenodo,
-        )
-
-    def query(self, release: OaWebRelease, **kwargs):
-        """Fetch the data for each table.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-        results = []
-
-        # ROR release date
-        ror_table_id = "ror"
-        ror_release_date = select_table_shard_dates(
-            project_id=self.input_project_id,
-            dataset_id=self.ror_dataset_id,
-            table_id=ror_table_id,
-            end_date=release.release_date,
-        )[0]
-        ror_sharded_table_id = bigquery_sharded_table_id(ror_table_id, ror_release_date)
-
-        for agg_table_id in self.table_ids:
-            # Aggregate release dates
-            agg_release_date = select_table_shard_dates(
-                project_id=self.input_project_id,
-                dataset_id=self.agg_dataset_id,
-                table_id=agg_table_id,
-                end_date=release.release_date,
-            )[0]
-            agg_sharded_table_id = bigquery_sharded_table_id(agg_table_id, agg_release_date)
-
-            # Fetch data
-            destination_uri = f"gs://{release.download_bucket}/{self.dag_id}/{release.release_id}/{agg_table_id}.jsonl"
-            success = bq_query_to_gcs(
-                query=QUERY.format(
-                    start_year=START_YEAR,
-                    project_id=self.input_project_id,
-                    agg_dataset_id=self.agg_dataset_id,
-                    agg_table_id=agg_sharded_table_id,
-                    ror_dataset_id=self.ror_dataset_id,
-                    ror_table_id=ror_sharded_table_id,
-                    settings_dataset_id=self.settings_dataset_id,
-                    country_table_id="country",
-                ),
-                project_id=self.output_project_id,
-                destination_uri=destination_uri,
-            )
-            results.append(success)
-
-        state = all(results)
-        if not state:
-            raise AirflowException("OaWebWorkflow.query failed")
-
-    def download(self, release: OaWebRelease, **kwargs):
-        """Download the queried data.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-
-        prefix = f"{self.dag_id}/{release.release_id}"
-        state = download_blobs_from_cloud_storage(
-            bucket_name=release.download_bucket, prefix=prefix, destination_path=release.download_folder
-        )
-        if not state:
-            raise AirflowException("OaWebWorkflow.download failed")
-
-    def make_draft_zenodo_version(self, release: OaWebRelease, **kwargs):
-        """Make a draft Zenodo version of the dataset.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-
-        make_draft_version(release.zenodo, self.conceptrecid)
-
-    def download_twitter_cards(self, release: OaWebRelease, **kwargs):
-        """Download current twitter cards
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-
-        # Download twitter cards
-        blob_name = "twitter.zip"
-        file_path = os.path.join(release.transform_folder, blob_name)
-        download_blob_from_cloud_storage(bucket_name=release.data_bucket_name, blob_name=blob_name, file_path=file_path)
-
-        # Unzip into build
-        unzip_folder_path = os.path.join(release.build_path)
-        with ZipFile(file_path) as zip_file:
-            zip_file.extractall(unzip_folder_path)
-
-    def transform(self, release: OaWebRelease, **kwargs):
-        """Transform the queried data into the final format for the open access website.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-
-        # Get versions
-        res = release.zenodo.get_versions(self.conceptrecid, all_versions=1)
-        if res.status_code != 200:
-            raise AirflowException(f"zenodo.get_versions status_code {res.status_code}")
-        versions = res.json()
-
-        # Make required folders
-        all_entities = []
-        countries = []
-        institutions = []
-        build_data_path = os.path.join(release.build_path, "data")
-        os.makedirs(build_data_path, exist_ok=True)
-        for category in self.table_ids:
-            logging.info(f"Transforming {category} entity")
-            # Load data
-            path = os.path.join(release.download_folder, f"{category}.jsonl")
-            df = self.load_data(path)
-
-            # Pre-process data
-            df = self.preprocess_df(category, df)
-
-            # Make index table
-            df_index_table = self.make_index(category, df)
-            self.update_index_with_logos(release.build_path, release.assets_path, category, df_index_table)
-            self.update_index_with_wiki_descriptions(df_index_table)
-            entities = self.make_entities(category, df_index_table, df)
-
-            # Make search index data for this category
-            all_entities += entities
-
-            # Save index
-            index_path = os.path.join(build_data_path, f"{category}.json")
-            self.save_index(index_path, entities)
-
-            # Save entities
-            entities_path = os.path.join(build_data_path, category)
-            self.save_entities(entities_path, entities)
-            logging.info(f"Saved transformed {category} entity")
-
-            # Assign country or institution variables
-            if category == "country":
-                countries = entities
-            elif category == "institution":
-                institutions = entities
-            else:
-                raise AirflowException(f"Category type unknown: {category}")
-
-        # Save all entities as json
-        index_path = os.path.join(build_data_path, f"index.json")
-        self.save_index(index_path, all_entities)
-
-        # Save COKI Open Access Dataset
-        coki_dataset_path = os.path.join(release.transform_folder, "coki-oa-dataset")
-        self.save_coki_oa_dataset(coki_dataset_path, countries, institutions)
-
-        # Make stats
-        end_year = pendulum.now().year - 1
-        zenodo_versions = [
-            ZenodoVersion(
-                pendulum.parse(version["created"]),
-                f"https://zenodo.org/record/{version['id']}/files/coki-oa-dataset.zip?download=1",
-            )
-            for version in versions
-        ]
-        last_updated = zenodo_versions[0].release_date.format("D MMMM YYYY")
-        country_stats = make_entity_stats(countries)
-        institution_stats = make_entity_stats(institutions)
-        stats = Stats(START_YEAR, end_year, last_updated, zenodo_versions, country_stats, institution_stats)
-        stats_path = os.path.join(release.build_path, "data")
-        self.save_stats(stats_path, stats)
-        logging.info(f"Saved stats data")
-
-        # Zip data
-        dst = os.path.join(release.transform_folder, "latest")
-        shutil.copytree(release.build_path, dst)
-        shutil.make_archive(dst, "zip", dst)
-
-    def publish_zenodo_version(self, release: OaWebRelease, **kwargs):
-        """Publish the new Zenodo version of the dataset.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-
-        zenodo = release.zenodo
-        res = zenodo.get_versions(self.conceptrecid, all_versions=0)
-        if res.status_code != 200:
-            raise AirflowException(f"zenodo.get_versions status_code {res.status_code}")
-        draft = res.json()[0]
-        draft_id = draft["id"]
-        if draft["state"] != "unsubmitted":
-            raise AirflowException(f"Latest version is not a draft: {draft_id}")
-
-        file_path = os.path.join(release.transform_folder, "coki-oa-dataset.zip")
-        publish_new_version(release.zenodo, draft_id, file_path)
-
-    def upload_dataset(self, release: OaWebRelease, **kwargs):
-        """Publish the dataset produced by this workflow.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-
-        # upload_file_to_cloud_storage should always rewrite a new version of latest.zip if it exists
-        # object versioning on the bucket will keep the previous versions
-        blob_name = f"{self.version}/latest.zip"
-        file_path = os.path.join(release.transform_folder, "latest.zip")
-        upload_file_to_cloud_storage(
-            bucket_name=release.data_bucket_name, blob_name=blob_name, file_path=file_path, check_blob_hash=False
-        )
-
-    def repository_dispatch(self, release: OaWebRelease, **kwargs):
-        """Trigger a Github repository_dispatch to trigger new website builds.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-
-        token = get_airflow_connection_password(self.GITHUB_TOKEN_CONN)
-        event_types = ["data-update/develop", "data-update/staging", "data-update/production"]
-        for event_type in event_types:
-            trigger_repository_dispatch(token=token, event_type=event_type)
-
-    def cleanup(self, release: OaWebRelease, **kwargs):
-        """Delete all files and folders associated with this release.
-
-        :param release: the release.
-        :param kwargs: the context passed from the PythonOperator. See
-        https://airflow.apache.org/docs/stable/macros-ref.html for a list of the keyword arguments that are
-        passed to this argument.
-        :return: None.
-        """
-        release.cleanup()
-
-    ######################################
-    # Other Functions
-    ######################################
-
-    def load_data(self, file_path: str) -> pd.DataFrame:
-        """Load the data file for a given category.
-
-        :param file_path: the path to the file to load.
-        :return: the Pandas Dataframe.
-        """
-
-        data = load_jsonl(file_path)
-        return pd.DataFrame(data)
-
-    def preprocess_df(self, category: str, df: pd.DataFrame) -> pd.DataFrame:
-        """Pre-process the data frame.
-
-        :param category: the category.
-        :param df: the dataframe.
-        :return: the Pandas Dataframe.
-        """
-
-        # Convert data types
-        df = df.copy(deep=True)
-        df["date"] = pd.to_datetime(df["date"])
-        df.fillna("", inplace=True)
-        for column in df.columns:
-            if column.startswith("n_"):
-                df[column] = pd.to_numeric(df[column])
-
-        # Clean RoR ids
-        if category == "institution":
-            # Remove columns not used for institutions
-            df.drop(columns=["alpha2"], inplace=True, errors="ignore")
-
-            # Clean RoR ids
-            df["id"] = df["id"].apply(lambda i: clean_ror_id(i))
-
-            # Parse identifiers
-            preferred_key = "preferred"
-            identifiers = []
-            for i, row in df.iterrows():
-                # Parse identifier for each entry
-                ent_ids = []
-                ids_dict = row["identifiers"]
-
-                # Add ROR id
-                ror_id = row["id"]
-                ent_ids.append({"id": ror_id, "type": "ROR", "url": f"https://ror.org/{ror_id}"})
-
-                # Parse other ids
-                for k, v in ids_dict.items():
-                    url = None
-                    id_type = k
-                    if id_type != "OrgRef":
-                        if preferred_key in v:
-                            id_value = v[preferred_key]
-                        else:
-                            id_value = v["all"][0]
-
-                        # Create URLs
-                        if id_type == "ISNI":
-                            url = f"https://isni.org/isni/{id_value}"
-                        elif id_type == "Wikidata":
-                            url = f"https://www.wikidata.org/wiki/{id_value}"
-                        elif id_type == "GRID":
-                            url = f"https://grid.ac/institutes/{id_value}"
-                        elif id_type == "FundRef":
-                            url = f"https://api.crossref.org/funders/{id_value}"
-
-                        ent_ids.append({"id": id_value, "type": id_type, "url": url})
-                identifiers.append(ent_ids)
-            df["identifiers"] = identifiers
-
-        if category == "country":
-            # Remove columns not used for countries
-            df.drop(columns=["url", "institution_types", "country", "identifiers"], inplace=True, errors="ignore")
-
-        return df
-
-    def make_index(self, category: str, df: pd.DataFrame):
-        """Make the data for the index tables.
-
-        :param category: the category, i.e. country or institution.
-        :param df: Pandas dataframe with all data points.
-        :return:
-        """
-
-        # Create aggregate
-        agg = {}
-        for column in df.columns:
-            if column.startswith("n_"):
-                agg[column] = "sum"
-            else:
-                agg[column] = "first"
-
-        # Create aggregate
-        df_index_table = df.groupby(["id"]).agg(
-            agg,
-            index=False,
-        )
-
-        # Exclude countries with small samples
-        df_index_table = df_index_table[df_index_table["n_outputs"] >= INCLUSION_THRESHOLD[category]]
-
-        # Add percentages to dataframe
-        self.update_df_with_percentages(df_index_table, self.PERCENTAGE_FIELD_KEYS)
-
-        # Sort from highest oa percentage to lowest
-        df_index_table.sort_values(by=["n_outputs_open"], ascending=False, inplace=True)
-
-        # Add category
-        df_index_table["category"] = category
-
-        # Remove date, year and repositories
-        df_index_table.drop(columns=["date", "year", "repositories"], inplace=True)
-
-        return df_index_table
-
-    def update_df_with_percentages(self, df: pd.DataFrame, keys: List[Tuple[str, str]]):
-        """Calculate percentages for fields in a Pandas dataframe.
-
-        :param df: the Pandas dataframe.
-        :param keys: they keys to calculate percentages for.
-        :return: None.
-        """
-
-        for numerator_key, denominator_key in keys:
-            p_key = f"p_{numerator_key}"
-            df[p_key] = round(df[f"n_{numerator_key}"] / df[denominator_key] * 100, 2)
-
-            # Fill in NaN caused by denominator of zero
-            df[p_key] = df[p_key].fillna(0)
-
-    def update_index_with_logos(self, build_path: str, assets_path: str, category: str, df_index_table: pd.DataFrame):
-        """Update the index with logos, downloading logos if they don't exist.
-
-        :param build_path: the path to the build folder.
-        :param assets_path: the path to oa-web-workflow assets folder which contains country flags.
-        :param category: the category, i.e. country or institution.
-        :param df_index_table: the index table Pandas dataframe.
-        :return: None.
-        """
-
-        sizes = ["s", "l", "xl"]
-        for size in sizes:
-            base_path = os.path.join(build_path, "logos", category, size)
-            os.makedirs(base_path, exist_ok=True)
-
-        # Make logos
-        if category == "country":
-            logging.info("Copying local logos")
-            country_sizes = sizes[:2]
-            with ThreadPoolExecutor() as executor:
-                futures = []
-                # Copy and rename logo images from using alpha2 to alpha3 country codes
-                for size in country_sizes:
-                    base_path = os.path.join(build_path, "logos", category, size)
-                    for alpha3, alpha2 in zip(df_index_table["id"], df_index_table["alpha2"]):
-                        src_path = os.path.join(assets_path, "flags", size, f"{alpha2}.svg")
-                        dst_path = os.path.join(base_path, f"{alpha3}.svg")
-                        futures.append(executor.submit(shutil.copy, src_path, dst_path))
-                [f.result() for f in as_completed(futures)]
-            logging.info("Finished copying local logos")
-
-            # Add logo urls to index
-            for size in sizes:
-                # For xl size point to l svg
-                make_logo_url_size = size
-                if size == "xl":
-                    make_logo_url_size = "l"
-                df_index_table[f"logo_{size}"] = df_index_table["id"].apply(
-                    lambda country_code: make_logo_url(
-                        category=category, entity_id=country_code, size=make_logo_url_size, fmt="svg"
-                    )
-                )
-
-        elif category == "institution":
-            # Get the institution logo and the path to the logo image
-            logging.info("Downloading logos using Clearbit")
-            institution_sizes = [("s", 32, "jpg"), ("l", 128, "jpg"), ("xl", 532, "png")]
-            for size, width, fmt in institution_sizes:
-                with ThreadPoolExecutor() as executor:
-                    futures = []
-                    logo_paths = []
-                    for ror_id, url in zip(df_index_table["id"], df_index_table["url"]):
-                        if url:
-                            url = clean_url(url)
-                            futures.append(
-                                executor.submit(get_institution_logo, ror_id, url, size, width, fmt, build_path)
-                            )
-                        else:
-                            logo_paths.append((ror_id, "/unknown.svg"))
-                    logo_paths += [f.result() for f in as_completed(futures)]
-                logging.info("Finished downloading logos")
-
-                # Sort table and results by id
-                df_index_table.sort_index(inplace=True)
-                logo_paths_sorted = [tup[1] for tup in sorted(logo_paths, key=lambda tup: tup[0])]
-
-                # Add logo paths to table
-                df_index_table[f"logo_{size}"] = logo_paths_sorted
-
-    def update_index_with_wiki_descriptions(self, df_index_table: pd.DataFrame):
-        """Get the wikipedia descriptions for each entity (institution or country) and add them to the index table.
-
-        :param df_index_table: the index table Pandas dataframe.
-        :return: None.
-        """
-        # Filter to select rows where url is not empty
-        wikipedia_url_filter = df_index_table["wikipedia_url"] != ""
-
-        # The wikipedia 'title' is the last part of the wikipedia url, without segments specified with '#'
-        titles_all = list(
-            zip(
-                df_index_table.loc[wikipedia_url_filter, "wikipedia_url"]
-                .str.split("wikipedia.org/wiki/")
-                .str[-1]
-                .str.split("#")
-                .str[0],
-                df_index_table.loc[wikipedia_url_filter, "id"],
-            )
-        )
-        # Create list with dictionaries of max 20 ids + titles (this is wiki api max)
-        titles_chunks = [
-            dict(titles_all[i : i + OaWebWorkflow.WIKI_MAX_TITLES])
-            for i in range(0, len(titles_all), OaWebWorkflow.WIKI_MAX_TITLES)
-        ]
-
-        logging.info(
-            f"Downloading wikipedia descriptions for all {len(titles_all)} entities in {len(titles_chunks)} chunks."
-        )
-        # Download 'punkt' resource, required when shortening wiki descriptions
-        nltk.download("punkt")
-
-        # Process each dictionary in separate thread to get wiki descriptions
+    return f"/logos/{category}/{size}/{entity_id}.{fmt}"
+
+
+def get_institution_logo(ror_id: str, url: str, size: str, width: int, fmt: str, build_path) -> Tuple[str, str]:
+    """Get the path to the logo for an institution.
+    If the logo does not exist in the build path yet, download from the Clearbit Logo API tool.
+    If the logo does not exist and failed to download, the path will default to "/unknown.svg".
+
+    :param ror_id: the institution's ROR id
+    :param url: the URL of the company domain + suffix e.g. spotify.com
+    :param size: the image size of the small logo for tables etc.
+    :param width: the width of the image.
+    :param fmt: the image format.
+    :param build_path: the build path for files of this workflow
+    :return: The ROR id and relative path (from build path) to the logo
+    """
+    logo_path = f"/unknown.svg"
+
+    file_path = os.path.join(build_path, "logos", "institution", size, f"{ror_id}.{fmt}")
+    if not os.path.isfile(file_path):
+        clearbit_download_logo(company_url=url, file_path=file_path, size=width, fmt=fmt)
+    if os.path.isfile(file_path):
+        logo_path = make_logo_url(category="institution", entity_id=ror_id, size=size, fmt=fmt)
+
+    return ror_id, logo_path
+
+
+def clean_url(url: str) -> str:
+    """Remove path and query from URL.
+
+    :param url: the url.
+    :return: the cleaned url.
+    """
+
+    p = urlparse(url)
+    return f"{p.scheme}://{p.netloc}/"
+
+
+def update_index_with_logos(build_path: str, assets_path: str, category: str, df_index_table: pd.DataFrame):
+    """Update the index with logos, downloading logos if they don't exist.
+
+    :param build_path: the path to the build folder.
+    :param assets_path: the path to oa-web-workflow assets folder which contains country flags.
+    :param category: the category, i.e. country or institution.
+    :param df_index_table: the index table Pandas dataframe.
+    :return: None.
+    """
+
+    sizes = ["s", "l", "xl"]
+    for size in sizes:
+        base_path = os.path.join(build_path, "logos", category, size)
+        os.makedirs(base_path, exist_ok=True)
+
+    # Make logos
+    if category == "country":
+        logging.info("Copying local logos")
+        country_sizes = sizes[:2]
         with ThreadPoolExecutor() as executor:
             futures = []
-            for titles in titles_chunks:
-                futures.append(executor.submit(get_wiki_descriptions, titles))
-            descriptions = []
-            for f in as_completed(futures):
-                descriptions += f.result()
-        logging.info(f"Finished downloading wikipedia descriptions")
+            # Copy and rename logo images from using alpha2 to alpha3 country codes
+            for size in country_sizes:
+                base_path = os.path.join(build_path, "logos", category, size)
+                for alpha3, alpha2 in zip(df_index_table["id"], df_index_table["alpha2"]):
+                    src_path = os.path.join(assets_path, "flags", size, f"{alpha2}.svg")
+                    dst_path = os.path.join(base_path, f"{alpha3}.svg")
+                    futures.append(executor.submit(shutil.copy, src_path, dst_path))
+            [f.result() for f in as_completed(futures)]
+        logging.info("Finished copying local logos")
 
-        # Sort table and results by id
-        df_index_table.sort_index(inplace=True)
-        descriptions_sorted = [tup[1] for tup in sorted(descriptions, key=lambda tup: tup[0])]
-
-        # Add wiki descriptions to table
-        df_index_table.loc[wikipedia_url_filter, "description"] = descriptions_sorted
-        df_index_table.loc[~wikipedia_url_filter, "description"] = ""
-
-    def save_index(self, file_path: str, entities: List[Entity]):
-        """Save an index file.
-
-        :param file_path: the file path where the index should be saved.
-        :param entities: a list of entities.
-        :return: None.
-        """
-
-        # Convert entities to dictionaries and select a subset of fields
-        subset = {
-            "id": None,
-            "name": None,
-            "logo_s": None,
-            "category": None,
-            "country": None,
-            "subregion": None,
-            "region": None,
-            "institution_types": None,
-            "acronyms": None,
-            "stats": {
-                "n_outputs": None,
-                "n_outputs_open": None,
-                "p_outputs_open": None,
-                "p_outputs_publisher_open_only": None,
-                "p_outputs_both": None,
-                "p_outputs_other_platform_open_only": None,
-                "p_outputs_closed": None,
-            },
-        }
-        data = [select_subset(entity.to_dict(), subset) for entity in entities]
-
-        # Save as JSON
-        save_json(file_path, data)
-
-    def make_entities(self, category: str, df_index_table: pd.DataFrame, df: pd.DataFrame) -> List[Entity]:
-        """Make entities.
-
-        :param category: the entity category.
-        :param df_index_table: the index table Pandas Dataframe.
-        :param df: the Pandas dataframe.
-        :return: the Entity objects.
-        """
-
-        entities = []
-        key_id = "id"
-        key_year = "year"
-        key_date = "date"
-        key_records = "records"
-        ts_groups = df.groupby([key_id])
-        for entity_id, df_group in ts_groups:
-            # Exclude institutions with small num outputs
-            total_outputs = df_group["n_outputs"].sum()
-            if total_outputs >= INCLUSION_THRESHOLD[category]:
-                self.update_df_with_percentages(df_group, self.PERCENTAGE_FIELD_KEYS)
-                df_group = df_group.sort_values(by=[key_year])
-                df_group = df_group.loc[:, ~df_group.columns.str.contains("^Unnamed")]
-
-                # Create entity
-                entity_dict: Dict = df_index_table.loc[df_index_table[key_id] == entity_id].to_dict(key_records)[0]
-                entity = Entity.from_dict(entity_dict)
-                entity.stats = PublicationStats.from_dict(entity_dict)
-
-                # Make timeseries data
-                years = []
-                rows: List[Dict] = df_group.to_dict(key_records)
-                for row in rows:
-                    year = int(row.get(key_year))
-                    date = row.get(key_date)
-                    stats = PublicationStats.from_dict(row)
-                    years.append(Year(year=year, date=date, stats=stats))
-                entity.years = years
-
-                # Make repositories
-                def sort_key(col):
-                    if is_string_dtype(col.dtype):
-                        return col.str.lower()
-                    return col
-
-                def merge_category(cat):
-                    if cat in {"Aggregator", "Unknown"}:
-                        return "Other Internet"
-                    return cat
-
-                repositories = []
-                for row in rows:
-                    repositories += row["repositories"]
-                df_repos = pd.DataFrame(repositories, columns=["id", "total_outputs", "category", "home_repo"])
-                df_repos["total_outputs"] = pd.to_numeric(df_repos["total_outputs"])
-                df_repos["category"] = df_repos["category"].apply(merge_category)
-                df_repos = (
-                    df_repos.groupby(["id"], as_index=False)
-                    .agg(
-                        {
-                            "id": "first",
-                            "total_outputs": "sum",
-                            "category": "first",
-                            "home_repo": "first",
-                        }
-                    )
-                    .sort_values(by=["total_outputs", "id"], ascending=[False, True], inplace=False, key=sort_key)
+        # Add logo urls to index
+        for size in sizes:
+            # For xl size point to l svg
+            make_logo_url_size = size
+            if size == "xl":
+                make_logo_url_size = "l"
+            df_index_table[f"logo_{size}"] = df_index_table["id"].apply(
+                lambda country_code: make_logo_url(
+                    category=category, entity_id=country_code, size=make_logo_url_size, fmt="svg"
                 )
-                repositories = df_repos.to_dict(key_records)
-                repositories = [Repository.from_dict(repo) for repo in repositories]
+            )
 
-                # Select a maximum number of repositories, however, make sure that all repositories that belong
-                # to the given institution are always present
-                entity.repositories = []
-                for repo in repositories:
-                    if len(entity.repositories) <= MAX_REPOSITORIES or repo.home_repo:
-                        entity.repositories.append(repo)
+    elif category == "institution":
+        # Get the institution logo and the path to the logo image
+        logging.info("Downloading logos using Clearbit")
+        institution_sizes = [("s", 32, "jpg"), ("l", 128, "jpg"), ("xl", 532, "png")]
+        for size, width, fmt in institution_sizes:
+            with ThreadPoolExecutor() as executor:
+                futures = []
+                logo_paths = []
+                for ror_id, url in zip(df_index_table["id"], df_index_table["url"]):
+                    if url:
+                        url = clean_url(url)
+                        futures.append(executor.submit(get_institution_logo, ror_id, url, size, width, fmt, build_path))
+                    else:
+                        logo_paths.append((ror_id, "/unknown.svg"))
+                logo_paths += [f.result() for f in as_completed(futures)]
+            logging.info("Finished downloading logos")
 
-                # Set min and max years for data
-                entity.start_year = years[0].year
-                entity.end_year = years[-1].year
+            # Sort table and results by id
+            df_index_table.sort_index(inplace=True)
+            logo_paths_sorted = [tup[1] for tup in sorted(logo_paths, key=lambda tup: tup[0])]
 
-                entities.append(entity)
+            # Add logo paths to table
+            df_index_table[f"logo_{size}"] = logo_paths_sorted
 
-        # Ensure that entities are sorted based on p_outputs_open
-        entities = sorted(entities, key=lambda e: e.stats.p_outputs_open, reverse=True)
 
-        return entities
+#########################
+# Wikipedia descriptions
+#########################
 
-    def save_entities(self, path: str, entities: List[Entity]):
-        """Save the data for each entity as a JSON file.
 
-        :param path: the path where the entities should be saved.
-        :param entities: the list of Entity objects.
-        :return: None.
-        """
+def update_index_with_wiki_descriptions(df_index_table: pd.DataFrame):
+    """Get the wikipedia descriptions for each entity (institution or country) and add them to the index table.
 
-        os.makedirs(path, exist_ok=True)
-        for entity in entities:
-            output_path = os.path.join(path, f"{entity.id}.json")
-            entity_dict = entity.to_dict()
-            save_json(output_path, entity_dict)
+    :param df_index_table: the index table Pandas dataframe.
+    :return: None.
+    """
+    # Filter to select rows where url is not empty
+    wikipedia_url_filter = df_index_table["wikipedia_url"] != ""
 
-    def save_coki_oa_dataset(self, path: str, countries: List[Entity], institutions: List[Entity]):
-        """Save the COKI Open Access Dataset to a zip file.
-
-        :param path: the path to the folder where the dataset should be saved.
-        :param countries: the country entities.
-        :param institutions: the institution entities.
-        :return: None.
-        """
-
-        # Country table
-        subset = {
-            "id": None,
-            "name": None,
-            "subregion": None,
-            "region": None,
-            "start_year": None,
-            "end_year": None,
-            "stats": None,
-            "years": None,
-        }
-        country = [select_subset(entity.to_dict(), subset) for entity in countries]
-
-        # Institutions table
-        subset = {
-            "id": None,
-            "name": None,
-            "country": None,
-            "subregion": None,
-            "region": None,
-            "institution_types": None,
-            "start_year": None,
-            "end_year": None,
-            "stats": None,
-            "years": None,
-        }
-        institution = [select_subset(entity.to_dict(), subset) for entity in institutions]
-
-        # Save to JSON Lines
-        os.makedirs(path, exist_ok=True)
-
-        file_path = os.path.join(path, "country.jsonl")
-        save_as_jsonl(file_path, country)
-
-        file_path = os.path.join(path, "institution.jsonl")
-        save_as_jsonl(file_path, institution)
-
-        # Save README
-        file_path = os.path.join(path, "README.md")
-        template = Template(README, keep_trailing_newline=True)
-        rendered = template.render(
-            year=pendulum.now().year, n_countries=len(countries), n_institutions=len(institutions)
+    # The wikipedia 'title' is the last part of the wikipedia url, without segments specified with '#'
+    titles_all = list(
+        zip(
+            df_index_table.loc[wikipedia_url_filter, "wikipedia_url"]
+            .str.split("wikipedia.org/wiki/")
+            .str[-1]
+            .str.split("#")
+            .str[0],
+            df_index_table.loc[wikipedia_url_filter, "id"],
         )
-        with open(file_path, mode="w") as f:
-            f.write(rendered)
+    )
+    # Create list with dictionaries of max 20 ids + titles (this is wiki api max)
+    titles_chunks = [dict(titles_all[i : i + WIKI_MAX_TITLES]) for i in range(0, len(titles_all), WIKI_MAX_TITLES)]
 
-        # Zip
-        shutil.make_archive(path, "zip", path)
+    logging.info(
+        f"Downloading wikipedia descriptions for all {len(titles_all)} entities in {len(titles_chunks)} chunks."
+    )
+    # Download 'punkt' resource, required when shortening wiki descriptions
+    nltk.download("punkt")
 
-    def save_stats(self, path: str, stats: Stats):
-        """Save overall stats.
+    # Process each dictionary in separate thread to get wiki descriptions
+    with ThreadPoolExecutor() as executor:
+        futures = []
+        for titles in titles_chunks:
+            futures.append(executor.submit(fetch_wiki_descriptions, titles))
+        descriptions = []
+        for f in as_completed(futures):
+            descriptions += f.result()
+    logging.info(f"Finished downloading wikipedia descriptions")
 
-        :param path: the directory where the stats.json file should be saved.
-        :param stats: stats object.
-        :return: None.
-        """
+    # Sort table and results by id
+    df_index_table.sort_index(inplace=True)
+    descriptions_sorted = [tup[1] for tup in sorted(descriptions, key=lambda tup: tup[0])]
 
-        os.makedirs(path, exist_ok=True)
+    # Add wiki descriptions to table
+    df_index_table.loc[wikipedia_url_filter, "description"] = descriptions_sorted
+    df_index_table.loc[~wikipedia_url_filter, "description"] = ""
 
-        # Save as JSON
-        output_path = os.path.join(path, "stats.json")
-        save_json(output_path, stats.to_dict())
+
+#############
+# Save data
+#############
+
+
+def save_entities(path: str, entities: List[Entity]):
+    """Save the data for each entity as a JSON file.
+
+    :param path: the path where the entities should be saved.
+    :param entities: the list of Entity objects.
+    :return: None.
+    """
+
+    os.makedirs(path, exist_ok=True)
+    for entity in entities:
+        output_path = os.path.join(path, f"{entity.id}.json")
+        entity_dict = entity.to_dict()
+        save_json(output_path, entity_dict)
+
+
+def save_stats(path: str, stats: Stats):
+    """Save overall stats.
+
+    :param path: the directory where the stats.json file should be saved.
+    :param stats: stats object.
+    :return: None.
+    """
+
+    os.makedirs(path, exist_ok=True)
+
+    # Save as JSON
+    output_path = os.path.join(path, "stats.json")
+    save_json(output_path, stats.to_dict())
+
+
+def save_coki_oa_dataset(path: str, countries: List[Entity], institutions: List[Entity]):
+    """Save the COKI Open Access Dataset to a zip file.
+
+    :param path: the path to the folder where the dataset should be saved.
+    :param countries: the country entities.
+    :param institutions: the institution entities.
+    :return: None.
+    """
+
+    # Country table
+    subset = {
+        "id": None,
+        "name": None,
+        "subregion": None,
+        "region": None,
+        "start_year": None,
+        "end_year": None,
+        "stats": None,
+        "years": None,
+    }
+    country = [select_subset(entity.to_dict(), subset) for entity in countries]
+
+    # Institutions table
+    subset = {
+        "id": None,
+        "name": None,
+        "country": None,
+        "subregion": None,
+        "region": None,
+        "institution_types": None,
+        "start_year": None,
+        "end_year": None,
+        "stats": None,
+        "years": None,
+    }
+    institution = [select_subset(entity.to_dict(), subset) for entity in institutions]
+
+    # Save to JSON Lines
+    os.makedirs(path, exist_ok=True)
+
+    file_path = os.path.join(path, "country.jsonl")
+    save_as_jsonl(file_path, country)
+
+    file_path = os.path.join(path, "institution.jsonl")
+    save_as_jsonl(file_path, institution)
+
+    # Save README
+    file_path = os.path.join(path, "README.md")
+    template = Template(README, keep_trailing_newline=True)
+    rendered = template.render(year=pendulum.now().year, n_countries=len(countries), n_institutions=len(institutions))
+    with open(file_path, mode="w") as f:
+        f.write(rendered)
+
+    # Zip
+    shutil.make_archive(path, "zip", path)
